+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `sourceMerchantId` to `WebhookNotification`s if present
+
 ## 2.5.0
 * Add support for Level 3 summary parameters: `shippingAmount`, `discountAmount`, and `shipsFromPostalCode`
 * Fix spec to expect PayPal transactions to move to settling rather than settled